% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMarkerSupport.R
\name{cellMarkerSupport}
\alias{cellMarkerSupport}
\title{Generate Marker Gene Evidence Matrix (CellMarker 2.0 DB)}
\usage{
cellMarkerSupport(
  markerDB,
  tissueClass = "all",
  tissueType = "all",
  species = "human",
  cellState = c("normal", "cancer"),
  markerSource = c("Experiment", "Review", "Single-cell sequencing", "Company"),
  logBase = exp(1),
  pseudoCount = 1
)
}
\arguments{
\item{markerDB}{A data frame containing the marker gene database.}

\item{tissueClass}{A character vector specifying the broad tissue classes to consider (case insensitive). Default is "all" that includes all tissue classes. To explore the available tissue classes, use the \code{showCellMarkerTissue()} function.}

\item{tissueType}{A character vector specifying the specific tissue types to consider within the chosen tissue class (case insensitive). Default is "all" that includes all tissue types. "tissueType" represents the specific subtypes within "tissueClass", For instance, when considering blood as "tissueClass", examples of blood tissue types could be "Peripheral blood", "Serum", "Blood vessel", and so on. To explore the available tissue types, use the \code{showCellMarkerTissue()} function.}

\item{species}{A character vector specifying the species to consider. Currently it supports "human" and "mouse". Use "all" to include all species.}

\item{cellState}{A character vector specifying the cell states to consider. Valid options are "normal" and "cancer". Default is both.}

\item{markerSource}{A character vector specifying the marker gene sources to consider. Valid options are "experiment", "review", "single-cell sequencing", and "company". Default is all.}

\item{logBase}{The base value for logarithmic transformation. Default is the Euler's number.}

\item{pseudoCount}{The pseudo count value added to the raw counts before logarithmic transformation. Default is 1.}
}
\value{
A matrix containing marker gene evidence levels. Row names represent cell types, and column names represent genes. The higher the value, the stronger the evidence that the gene is a marker for the cell type.
}
\description{
This internal function generates a marker gene evidence matrix based on the CellMarker 2.0 database and given filtering criteria.
}
