% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTaxonomySupport.R
\name{cellTaxonomySupport}
\alias{cellTaxonomySupport}
\title{Generate Marker Gene Evidence Matrix (Cell Taxonomy DB)}
\usage{
cellTaxonomySupport(
  markerDB,
  species = "Homo sapiens",
  logBase = exp(1),
  pseudoCount = 1
)
}
\arguments{
\item{markerDB}{A data frame containing the marker gene database.}

\item{species}{A character vector specifying the species to consider. Currently it contains data on 34 species. Use "all" to include all species. Default is "Homo sapiens". To explore the available species, use the \code{showCellTaxoSpecies()} function.}

\item{logBase}{The base value for logarithmic transformation. Default is the Euler's number.}

\item{pseudoCount}{The pseudo count value added to the raw counts before logarithmic transformation. Default is 1.}
}
\value{
A matrix containing marker gene evidence levels. Row names represent cell types, and column names represent genes. The higher the value, the stronger the evidence that the gene is a marker for the cell type.
}
\description{
This internal function generates a marker gene evidence matrix based on the Cell Taxonomy database and given filtering criteria.
}
