% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTables.R
\name{showCellMarkerTissue}
\alias{showCellMarkerTissue}
\title{Show CellMarker Database Tissue Information}
\usage{
showCellMarkerTissue()
}
\value{
an HTML widget to display tabular data.
}
\description{
This function displays a searchable table with the tissue information and the corresponding species from the CellMarker 2.0 database.
}
\examples{
showCellMarkerTissue()

}
