% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTables.R
\name{showCellTaxoSpecies}
\alias{showCellTaxoSpecies}
\title{Show Cell Taxonomy Database Species Information}
\usage{
showCellTaxoSpecies()
}
\value{
an HTML widget to display tabular data.
}
\description{
This function displays a searchable table with the species information and the number of corresponding cell types from the Cell Taxonomy database.
}
\examples{
showCellTaxoSpecies()

}
