% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCellTypeRes.R
\name{saveCellTypeRes}
\alias{saveCellTypeRes}
\title{Save Cell Type Results}
\usage{
saveCellTypeRes(cellTypeRes, path)
}
\arguments{
\item{cellTypeRes}{A list containing the annotation results for each cell cluster.}

\item{path}{The path where the Excel file will be saved. For example, path = "directory/results.xlsx".}
}
\value{
None
}
\description{
This function saves the cell type results to an Excel file.
}
